better file structure + added some comments about the implementation
